Extracting Trace lineages functionality
